refactor: add relative paths
