+++ package.json
@@ -11,6 +11,7 @@
     "mobx-react": "6.1.4",
     "mobx-state-tree": "3.15.0",
     "mst-middlewares": "3.15.0",
+    "prop-types": "^15.8.1",
     "react": "16.12.0",
     "react-dom": "16.12.0",
     "react-scripts": "^5.0.1",
