+++ src/components/index.js
@@ -1,0 +1,4 @@
+export * from './Canvas/Canvas'
+export * from './Box/Box'
+export * from './Box/BoxDraggable'
+export * from './Toolbar/Toolbar'
