+++ src/stores/index.js
@@ -1,0 +1,3 @@
+export * from './models/Box'
+export * from './models/CursorPointer'
+export * from './models/MainStore'
