+++ src/stores/utils/createBox.js
@@ -1,13 +1,11 @@
-
-import uuid from 'uuid/v4';
-import getRandomColor from '../../utils/getRandomColor'
+import { getRandomColor, getRandomUuid } from '../../utils'
 import { BoxModel } from '../models/Box'
 
 export const createBox = (left, top) => {
   return BoxModel.create({
-    id: uuid(),
+    id: getRandomUuid(),
     color: getRandomColor(),
     left,
-    top
-  });
-};
+    top,
+  })
+}
