+++ src/utils/getRandomColor.js
@@ -1,4 +1,4 @@
-const getRandomColor = () => {
+export const getRandomColor = () => {
   const letters = '0123456789ABCDEF'
   let color = '#'
   for (let i = 0; i < 6; i++) {
@@ -6,5 +6,3 @@ const getRandomColor = () => {
   }
   return color
 }
-
-export default getRandomColor
