+++ src/utils/getRandomUuid.js
@@ -1,0 +1,5 @@
+import uuid from 'uuid/v4'
+
+export const getRandomUuid = () => {
+  return uuid()
+}
