+++ src/utils/index.js
@@ -1,0 +1,3 @@
+export * from './getRandomColor'
+export * from './getRandomUuid'
+export * from './getCursorPosition'
